Added shapes to bodies

+++ BasicSimulation/src/Position.h
@@ -2,18 +2,18 @@
 #define POSITION_H
 
 struct Position {
-    double sfCenterX;
-    double sfCenterY;
-    double sfOrientation;
+    float sfCenterX;
+    float sfCenterY;
+    float sfOrientation;
 
-    Position(double pCenterX, double pCenterY, double pOrientation) :
+    Position(float pCenterX, float pCenterY, float pOrientation) :
     sfCenterX(pCenterX),
     sfCenterY(pCenterY),
     sfOrientation(pOrientation) {}
 
-    double getCenterX() const { return sfCenterX; }
-    double getCenterY() const { return sfCenterY; }
-    double getOrientation() const { return sfOrientation; }
+    float getCenterX() const { return sfCenterX; }
+    float getCenterY() const { return sfCenterY; }
+    float getOrientation() const { return sfOrientation; }
 };
 
 #endif //POSITION_H
