+++ BasicSimulation/src/Velocity.h
@@ -2,21 +2,21 @@
 #define VELOCITY_H
 
 struct Velocity {
-    double sfSpeedX;
-    double sfSpeedY;
-    double sfAccelerationX;
-    double sfAccelerationY;
+    float sfSpeedX;
+    float sfSpeedY;
+    float sfAccelerationX;
+    float sfAccelerationY;
 
-    Velocity(double pSpeedX, double pSpeedY, double pAccelerationX, double pAccelerationY) :
+    Velocity(float pSpeedX, float pSpeedY, float pAccelerationX, float pAccelerationY) :
     sfSpeedX(pSpeedX),
     sfSpeedY(pSpeedY),
     sfAccelerationX(pAccelerationX),
     sfAccelerationY(pAccelerationY) {}
 
-    double getSpeedX() const { return sfSpeedX; }
-    double getSpeedY() const { return sfSpeedY; }
-    double getAccelerationX() const { return sfAccelerationX; }
-    double getAccelerationY() const { return sfAccelerationY; }
+    float getSpeedX() const { return sfSpeedX; }
+    float getSpeedY() const { return sfSpeedY; }
+    float getAccelerationX() const { return sfAccelerationX; }
+    float getAccelerationY() const { return sfAccelerationY; }
 };
 
 #endif //VELOCITY_H
